pyproject.toml
[project]
name = "baezi"
version = "0.1.0"
description = "Banking4 zu ezbookkeeping Importer"
license = {text = "MIT"}
keywords = ["banking", "ezbookkeeping", "import", "finance"]
readme = "README.md"
authors = [{name = "CM000n"}]
requires-python = ">=3.12,<3.15"
dependencies = [
    "requests>=2.32",
    "rich-click>=1.9",
    "python-dotenv>=1.0",
]

[project.scripts]
baezi = "baezi.__main__:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pre-commit>=4.0",
    "mypy>=1.0",
    "ruff>=0.8",
]

[tool.poetry]
version = "0.1.0"
package-mode = false
requires-poetry = ">=2.0,<3.0"

[tool.poetry.dependencies]
python = ">=3.12,<3.15"
requests = "^2.32"
rich-click = "^1.9"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0"
pytest = "^8.0"
pytest-cov = "^5.0"
mypy = "^1.0"
ruff = "^0.8"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
fix = true
line-length = 100
target-version = "py312"

[tool.ruff.lint]
exclude = ["tests/**"]
select = ["ALL"]
extend-ignore = ["D203", "D213", "COM812", "ISC001"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
split-on-trailing-comma = false

[tool.ruff.format]
docstring-code-format = true
skip-magic-trailing-comma = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--cov=baezi",
    "--cov-report=html",
    "--cov-report=term-missing",
    "-v",
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
